# util/decimal: Implemented Decimal Ln, Log, and LogN using new convergence loop structure