These utility functions, along with the looping structure they use were
adapted from Rob Pike's Ivy project. With them, I was able to add native
decimal support for sql's LN and LOG commands, instead of deferring to
imprecise float computation and potentially losing information.

The use of the new looping structure, which iterates until computational
convergence to a specified precision, also improved the performance of
our previous decimal square root implementation.

```
name           old time/op  new time/op  delta
DecimalSqrt-4  52.2µs ± 0%  22.4µs ±10%  -57.16%  (p=0.016 n=4+5)
```

<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.svg" height="40" alt="Review on Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/4108)
<!-- Reviewable:end -->