@petermattis 2016-02-03
First pass comments. I'll need to complete this review later.

---

Review status: 0 of 6 files reviewed at latest revision, 3 unresolved discussions.

---

*[util/decimal/const.go, line 33 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4108#-K9cHLbtZDwGk2MzLztT:-K9cHLbtZDwGk2MzLztU:-242865917) ([raw file](https://github.com/cockroachdb/cockroach/blob/2e9f0b98bad62050dc505c0a36c5ec94874388ad/util/decimal/const.go#L33)):*
Looks like `decimalE` and `decimalPi` are unused. 

---

*[util/decimal/loop.go, line 37 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4108#-K9cGjaTGMn_jcd7iOn9:-K9cGjaTGMn_jcd7iOnA:-1478410572) ([raw file](https://github.com/cockroachdb/cockroach/blob/2e9f0b98bad62050dc505c0a36c5ec94874388ad/util/decimal/loop.go#L37)):*
Do these need to be pointers?

---

*[util/decimal/loop.go, line 52 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4108#-K9cGz_VflXZ5ssVrXAh:-K9cGz_VflXZ5ssVrXAi:-1024399102) ([raw file](https://github.com/cockroachdb/cockroach/blob/2e9f0b98bad62050dc505c0a36c5ec94874388ad/util/decimal/loop.go#L52)):*
You're counting digits before the decimal point, not bits here, right?

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4108#-:-K9cHW5muTjW5B4z0lYC:-658219438)*
<!-- Sent from Reviewable.io -->