@petermattis 2016-02-03
LGTM

---

Review status: 0 of 6 files reviewed at latest revision, 5 unresolved discussions.

---

*[util/decimal/const.go, line 54 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4108#-K9cfBg3w3XftBJ7lFLv:-K9cfBg3w3XftBJ7lFLw:-1790495045) ([raw file](https://github.com/cockroachdb/cockroach/blob/74fa6e552fee336d18464527bf15db67545c071c/util/decimal/const.go#L54)):*
I think this can be commented out as well.

---

*[util/decimal/loop.go, line 56 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4108#-K9cgJGj3HelIfb1IyNd:-K9cgJGj3HelIfb1IyNe:-328446957) ([raw file](https://github.com/cockroachdb/cockroach/blob/74fa6e552fee336d18464527bf15db67545c071c/util/decimal/loop.go#L56)):*
Wondering if you could do this calculation more directly using the decimal exponent and the number of bits in the mantissa.

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4108#-:-K9chynQ8h94aodkHlqB:676834233)*
<!-- Sent from Reviewable.io -->