@nvanbenschoten 2016-02-03
Review status: 0 of 6 files reviewed at latest revision, 5 unresolved discussions.

---

*[util/decimal/const.go, line 33 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4108#-K9cHLbtZDwGk2MzLztT:-K9cQoBptHRah6_FlLaD:-1106600699) ([raw file](https://github.com/cockroachdb/cockroach/blob/2e9f0b98bad62050dc505c0a36c5ec94874388ad/util/decimal/const.go#L33)):*
Commented out until a use case comes up.

---

*[util/decimal/const.go, line 54 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4108#-K9cfBg3w3XftBJ7lFLv:-K9dP8Dwzz6em79eglEw:-611590598) ([raw file](https://github.com/cockroachdb/cockroach/blob/74fa6e552fee336d18464527bf15db67545c071c/util/decimal/const.go#L54)):*
Used in tests. See `TestDecimalLogN`.

---

*[util/decimal/loop.go, line 37 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4108#-K9cGjaTGMn_jcd7iOn9:-K9cQZDGONACt7ZtnPHt:-499146009) ([raw file](https://github.com/cockroachdb/cockroach/blob/2e9f0b98bad62050dc505c0a36c5ec94874388ad/util/decimal/loop.go#L37)):*
Done.

---

*[util/decimal/loop.go, line 52 \[r1\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4108#-K9cGz_VflXZ5ssVrXAh:-K9cUweIj5CplqGSVasj:-404161676) ([raw file](https://github.com/cockroachdb/cockroach/blob/2e9f0b98bad62050dc505c0a36c5ec94874388ad/util/decimal/loop.go#L52)):*
Right, it should have been scaled by `1/Log10(2)`. Fixed.

---

*[util/decimal/loop.go, line 56 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/4108#-K9cgJGj3HelIfb1IyNd:-K9dQ_FAsc5cwd3tUgoK:-960597166) ([raw file](https://github.com/cockroachdb/cockroach/blob/74fa6e552fee336d18464527bf15db67545c071c/util/decimal/loop.go#L56)):*
Ah right! I forgot about that method. Done.

---


*Comments from the [review on Reviewable.io](https://reviewable.io:443/reviews/cockroachdb/cockroach/4108)*
<!-- Sent from Reviewable.io -->